Fix bug in printk
Previously when handling format strings we'd decrement the format string
length by 2.  This is not only wrong but kind of a weird way to do it.
Instead of decrementing the size variable, we just increment our loop
variable an extra time which makes a little bit more sense and works
correctly.